style : TimeConverter로 함수 뺌
